# src/main/resources/openapi/entsoe-price-service-api-v1.yml
openapi: 3.0.2
info:
  title: ENTSO-E - Price Service
  version: 1.0.0
  description: >-
    ##  ENTSO-E Service
    
    
    This service is made with ther purpose of simplifying the current API that is provided by  ENTSO-E,
    and also converting the responses into JSON format instead of the current XML format provided by
    ENTSO-E.
    
    
    This API has been developed for Grønnere Elforbrug, by Kenneth Thomsen.
  contact:
    name: Kenneth Thomsen
    url: 'https://elforbrug.nu'
    email: kontakt@kthomsen.dk
  license:
    name: GNU GPLv3
    url: 'https://www.gnu.org/licenses/gpl.txt'
servers:
  -
    url: 'https://entsoe.systemedz.com/api/{version}'
    description: Production environment - currently hosted by Systemedz (Kenneth Thomsen)
    variables:
      version:
        enum:
          - v1
        default: v1
        description: The API Version
  -
    url: 'https://localhost:8443/api/{version}'
    description: Localhost Environment (for testing)
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
        description: The API Version
paths:
...
  '/prices/{areaCode}/by-month/{month}':
    summary: Search Prices from ENTSO-E by a specific month.
    description: >-
      By providing an area code and a month, you can look up prices from ENTSO-E in the specified area
      in that entire month, just by specifying the month as a number between 1 and 12. So if you want to
      look up prices in June, the month would be 6.
      
      
      If you want to look up June 2020, you can add the ?year=2020 parameter to the end of the query
      (e.g. /prices/FR/by-month/6?year=2020).
    get:
      tags:
        - PriceSpecificController
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
          description: Return prices for the requested period.
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      operationId: getPricesByMonth
      summary: Get ENTSO-E Market Prices by a specific month (and year).
      description: |-
        ## Interval by Month
        __Interval: June of this year__
        * /prices/FR/by-month/6
        
        __Interval: June of 2020__
        * /prices/FR/by-month/6?year=2020
    parameters:
      -
        name: areaCode
        description: >-
          The Area Code represents the area of which the data comes from. This could be e.g. DK1
          which equals Denmark (West), or FR which is France.
        schema:
          $ref: '#/components/schemas/AreaCode'
        in: path
        required: true
      -
        name: month
        description: 'The month of the year you want to look up the prices for, e.g. 7 july.'
        schema:
          type: integer
        in: path
        required: true
      -
        name: year
        description: 'Specify a year, if you want to look up a month that is not in the current year, e.g. 2020.'
        schema:
          type: integer
        in: query
      -
        name: entsoe-security-token
        description: Security Token provided by ENTSO-E.
        schema:
          type: string
        in: header
        required: true
  '/prices/{areaCode}/by-week/{week}':
    summary: Search Prices from ENTSO-E by a specific week.
    description: >-
      By providing an area code and a week, you can look up prices from ENTSO-E in the specified area in
      that entire week, just by specifying the week as a number between 1 and 53.
      
      
      If you want to look up week 20 of 2020, you can add the ?year=2020 parameter to the end of the
      query (e.g. /prices/FR/by-week/20?year=2020).
    get:
      tags:
        - PriceSpecificController
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
          description: Return prices for the requested period.
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      operationId: getPricesByWeek
      summary: Get ENTSO-E Market Prices by a specific week (and year).
      description: |-
        ## Interval by Week
        __Interval: Week 36 of this year__
        * /prices/FR/by-week/36
        
        __Interval: Week 32 of 2020__
        * /prices/FR/by-week/32?year=2020
    parameters:
      -
        name: areaCode
        description: >-
          The Area Code represents the area of which the data comes from. This could be e.g. DK1
          which equals Denmark (West), or FR which is France.
        schema:
          $ref: '#/components/schemas/AreaCode'
        in: path
        required: true
      -
        name: year
        description: 'Specify a year, if you want to look up a week that is not in the current year, e.g. 2020.'
        schema:
          type: integer
        in: query
      -
        name: entsoe-security-token
        description: Security Token provided by ENTSO-E.
        schema:
          type: string
        in: header
        required: true
      -
        name: week
        description: 'The week of the year you want to look up the prices for, e.g. 20.'
        schema:
          type: integer
        in: path
        required: true
  '/prices/{areaCode}/by-year/{year}':
    summary: Search Prices from ENTSO-E by a specific year.
    description: >-
      By providing an area code and a year, you can look up prices from ENTSO-E in the specified area in
      that entire year, just by specifying the year as a number.
    get:
      tags:
        - PriceSpecificController
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
          description: Return prices for the requested period.
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      operationId: getPricesByYear
      summary: Get ENTSO-E Market Prices by a specific year.
    parameters:
      -
        name: areaCode
        description: >-
          The Area Code represents the area of which the data comes from. This could be e.g. DK1
          which equals Denmark (West), or FR which is France.
        schema:
          $ref: '#/components/schemas/AreaCode'
        in: path
        required: true
      -
        name: entsoe-security-token
        description: Security Token provided by ENTSO-E.
        schema:
          type: string
        in: header
        required: true
      -
        name: year
        description: 'The year you want to look up the prices for, e.g. 2020.'
        schema:
          type: integer
        in: path
        required: true
  '/prices/{areaCode}/by-month/':
    summary: Search Prices from ENTSO-E from the current month.
    description: >-
      By providing an area code, you can look up prices from ENTSO-E in the specified area in the
      current month. If you want to look up another month, you can do that just by specifying the month
      as a number between 1 and 12 at the end of the URL path, e.g. /prices/FR/by-month/5 to get May of
      this year.
      
      
      If you want to look up May of 2020, you can add the ?year=2020 parameter to the end of the query
      (e.g. /prices/FR/by-month/5?year=2020).
    get:
      tags:
        - PriceSpecificController
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
          description: Return prices for the requested period.
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      operationId: getPricesCurrentMonth
      summary: Get ENTSO-E Market Prices for the current month.
      description: |-
        ## Interval by Month
        __Interval: The current month__
        * /prices/FR/by-month/
        
        __Interval: June of 2020__
        * /prices/FR/by-month/6?year=2020
    parameters:
      -
        name: areaCode
        description: >-
          The Area Code represents the area of which the data comes from. This could be e.g. DK1
          which equals Denmark (West), or FR which is France.
        schema:
          $ref: '#/components/schemas/AreaCode'
        in: path
        required: true
      -
        name: year
        description: 'Specify a year, if you want to look up a month that is not in the current year, e.g. 2020.'
        schema:
          type: integer
        in: query
      -
        name: entsoe-security-token
        description: Security Token provided by ENTSO-E.
        schema:
          type: string
        in: header
        required: true
  '/prices/{areaCode}/by-week/':
    summary: Search Prices from ENTSO-E from the current week.
    description: >-
      By providing an area code, you can look up prices from ENTSO-E in the specified area in the
      current week. If you want to look up another week, you can do that just by specifying the week as
      a number between 1 and 53 at the end of the URL path, e.g. /prices/FR/by-week/5 to get the 5th
      week of this year.
      
      
      If you want to look up week 32 of 2020, you can add the ?year=2020 parameter to the end of the
      query (e.g. /prices/FR/by-week/32?year=2020).
    get:
      tags:
        - PriceSpecificController
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
          description: Return prices for the requested period.
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      operationId: getPricesCurrentWeek
      summary: Get ENTSO-E Market Prices for the current month.
      description: |-
        ## Interval by Week
        __Interval: The current week__
        * /prices/FR/by-week/
        
        __Interval: Week 32 of 2020__
        * /prices/FR/by-week/32?year=2020
    parameters:
      -
        name: areaCode
        description: >-
          The Area Code represents the area of which the data comes from. This could be e.g. DK1
          which equals Denmark (West), or FR which is France.
        schema:
          $ref: '#/components/schemas/AreaCode'
        in: path
        required: true
      -
        name: year
        description: 'Specify a year, if you want to look up a week that is not in the current year, e.g. 2020.'
        schema:
          type: integer
        in: query
      -
        name: entsoe-security-token
        description: Security Token provided by ENTSO-E.
        schema:
          type: string
        in: header
        required: true
  '/prices/{areaCode}/by-year/':
    summary: Search Prices from ENTSO-E from the current year.
    description: >-
      By providing an area code, you can look up prices from ENTSO-E in the specified area in the
      current year. If you want to look up another year, you can do that just by specifying the year as
      a number at the end of the URL path, e.g. /prices/FR/by-year/2020 to get the prices from 2020.
    get:
      tags:
        - PriceSpecificController
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesResponse'
          description: Return prices for the requested period.
        '400':
          $ref: '#/components/responses/400BadRequest'
        '500':
          $ref: '#/components/responses/500InternalServerError'
      operationId: getPricesCurrentYear
      summary: Get ENTSO-E Market Prices for the current year.
      description: |-
        ## Interval by Year
        __Interval: The current year__
        * /prices/FR/by-year/
        
        __Interval: Year 2020__
        * /prices/FR/by-year/2020
    parameters:
      -
        name: areaCode
        description: >-
          The Area Code represents the area of which the data comes from. This could be e.g. DK1
          which equals Denmark (West), or FR which is France.
        schema:
          $ref: '#/components/schemas/AreaCode'
        in: path
        required: true
      -
        name: entsoe-security-token
        description: Security Token provided by ENTSO-E.
        schema:
          type: string
        in: header
        required: true
...
components:
  schemas:
    ErrorMessage:
      description: Error Message
      type: object
      properties:
        message:
          description: Error Message
          type: string
        details:
          description: ''
          type: array
          items:
            $ref: '#/components/schemas/ErrorMessageDetail'
      x-class-extra-annotation: >-
        @lombok.Builder
        @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL)
        @com.fasterxml.jackson.annotation.JsonPropertyOrder({ "message", "details" })
    ErrorMessageDetail:
      description: ''
      type: object
      properties:
        field:
          description: Field name
          type: string
        message:
          description: Information regarding the field error.
          type: string
      x-class-extra-annotation: >-
        @lombok.Builder
        @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.ALWAYS)
        @com.fasterxml.jackson.annotation.JsonPropertyOrder({ "field", "message" })
...
    AreaCode:
      description: |-
        Area Codes.
        * FR = France _(10YFR-RTE------C)_
        * DK1 = Denmark - West _(10YDK-1--------W)_
        * DK2 = Denmark - East _(10YDK-2--------M)_
        * DE = Germany _(10Y1001A1001A82H)_
        * SE1 = Sweden 1 _(10Y1001A1001A44P)_
        * SE2 = Sweden 2 _(10Y1001A1001A45N)_
        * SE3 = Sweden 3 _(10Y1001A1001A46L)_
        * SE4 = Sweden 4 _(10Y1001A1001A47J)_
        * NO1 = Norway 1 _(10YNO-1--------2)_
        * NO2 = Norway 2 _(10YNO-2--------T)_
        * NO3 = Norway 3 _(10YNO-3--------J)_
        * NO4 = Norway 4 _(10YNO-4--------9)_
        * NO5 = Norway 5 _(10Y1001A1001A48H)_
      enum:
        - FR
        - DK1
        - DK2
        - DE
        - SE1
        - SE2
        - SW1
        - SW2
        - SW3
        - SW4
        - NO1
        - NO2
        - NO3
        - NO4
        - NO5
        - FI
      type: string
    PricePoint:
      description: ''
      type: object
      properties:
        price:
          format: double
          description: ''
          type: number
        hour:
          description: ''
          type: string
      x-class-extra-annotation: >-
        @lombok.Builder
        @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.ALWAYS)
        @com.fasterxml.jackson.annotation.JsonPropertyOrder({ "hour", "price" })
    TimeInterval:
      description: ''
      required: []
      type: object
      properties:
        start:
          description: Interval Start Date/Time
          type: string
          example: '2022-10-15T22:00Z'
        end:
          description: Interval End Date/Time
          type: string
          example: '2022-10-17T22:00Z'
      x-class-extra-annotation: >-
        @lombok.Builder
        @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.ALWAYS)
        @com.fasterxml.jackson.annotation.JsonPropertyOrder({ "start", "end" })
    PricesResponse:
      description: ''
      required: []
      type: object
      properties:
        createdDateTime:
          description: Timestamp for Price Request creation.
          type: string
          example: '2022-10-19T05:00:55.466364'
        area:
          description: Area Name
          type: string
          example: FR - France
        areaCode:
          description: The ENTSO-E Area Code
          type: string
          example: 10YFR-RTE------C
        timeInterval:
          $ref: '#/components/schemas/TimeInterval'
          description: The Interval of the prices requested
        intervalDays:
          description: A list of the daily price information.
          type: array
          items:
            $ref: '#/components/schemas/IntervalDay'
      x-class-extra-annotation: >-
        @lombok.Builder
        @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.ALWAYS)
        @com.fasterxml.jackson.annotation.JsonPropertyOrder({ "area", "areaCode", "timeInterval",
        "intervalDays", "createdDateTime" })
    IntervalDay:
      description: The IntervalDay represents 1 day in the interval/range of the prices
      type: object
      properties:
        currency:
          description: Price Currency
          type: string
          example: EUR
        measureUnit:
          description: 'The measurement Unit related to the price, e.g. MWH.'
          type: string
          example: MWH
        hourPrices:
          description: A list of hourly prices throughout this Interval Date.
          maxItems: 24
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/PricePoint'
        currentDateInterval:
          $ref: '#/components/schemas/TimeInterval'
          description: The Interval of the current date
      x-class-extra-annotation: >-
        @lombok.Builder
        @com.fasterxml.jackson.annotation.JsonInclude(com.fasterxml.jackson.annotation.JsonInclude.Include.ALWAYS)
        @com.fasterxml.jackson.annotation.JsonPropertyOrder({ "currentDateInterval", "currency",
        "measureUnit", "hourPrices" })
  responses:
    400BadRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
          examples:
            Invalid Query Paramter:
              value:
                message: 'Invalid Query Parameter: interval=test'
                details:
                  -
                    field: interval
                    message: >-
                      Field should contain a value of type 'Integer'. Examples: 0, 1,
                      123, -1, -20 etc.
            Invalid Month Interval:
              value:
                message: Invalid parameters provided. See Details for more information.
                details:
                  -
                    field: interval
                    message: >-
                      Monthly Intervals are limited to +/- 12 months (1 year). If you
                      want to search for months further back in time, please use from/to
                      or year/month parameters instead, e.g.
                      /prices/FR?year=2020&month=10, or
                      /prices/FR?from=20201001&to=20201231
      description: >-
        When receiving a Bad Request Error Response, it means that the input provided in the API call
        is invalid.
        
        That could be an invalid combination of query paramters, an invalid ENTSO-E Security Token or
        invalid input in the API Request paramteres.
    500InternalServerError:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
          examples:
            Internal Error:
              value:
                message: An error occurred and has been reported to developers.
      description: >-
        Error Message Response when an Internal Server error happened, e.g. ENTSO-E does not respond
        etc.
        
        This error is not related to the request input, which means that the request is valid.
...